SimpleCard and Modal component moved to share folder, deleted components created as default in the project
